Refactor code, and add wchar_t and std::wstring support
IMPORTANT: This commit changes behaviour from parsing octal numbers.
Before, any scalar that consisted of numbers starting with a '0' was
considered octal, precisely like the functions std::stoi, std::stol,
std::stoll do. The YAML spec is not clear on this, but it looks like
a valid octal number starts with the prefix "0o" instead of just the
prefix "0". This commit fixes that.

IMPORTANT: This commit changes behaviour with respect to encoding
and decoding (un)signed integers. Observe that the values

    std::is_same<char, signed char>::value
    std::is_same<char, unsigned char>::value

are **always false**. What this means is that `char` is supposed to
be the general-purpose byte character type, suitable for strings.
Contrast this to (un)signed char, which is supposed to be an 8-bit
signed or unsigned number. Thus, emitting a signed char emits the
actual number that that signed char represents. Same holds for
unsigned.

Next, wchar_t and std::wstring support are added. They are converted
to and from multibyte UTF8 std::strings.

Next, the various converters are split up in separate files.

The YAML::Emitter has support for nearly all std containers.

Some decoder functions from primitive types have been changed from
using std::stringstream to using C++11's new family of string
conversion utility functions like std::stoi, std::stol, std::stoll
and so forth.